ECOSENT-1692 - ECOS process API for scenario editor

+++ src/test/java/ru/citeck/ecos/process/domain/impl/TimerServiceImplTest.java
@@ -116,7 +116,7 @@ public class TimerServiceImplTest {
         TimerEntity entity = entities.get(0);
 
         assertNotNull(entity.getId());
-        assertEquals((Integer) tenant, entity.getId().getTnt());
+        assertEquals((Integer) tenant, (Integer) entity.getId().getTnt());
         assertEquals(saved.getTimerId(), entity.getId().getId().toString());
 
         assertEquals(triggerTime, entity.getTriggerTime());
